fixed some minor problems with the intiallization scripts

+++ initialization/install_vimrc.sh
@@ -2,10 +2,21 @@
 
 # Since we are storing .vimrc, .gvimrc, and .vsvimrc in the ~/.vim/ directory
 # we need to make a symbolic link to these in the ~ directory
+if [ ! -f ~/.vimrc ]
+then
 ln -s ~/.vim/.vimrc ~/.vimrc
-ln -s ~/.vim/.gvimrc ~/.gvimrc
-ln -s ~/.vim/.vsvimrc ~/.vsvimrc
+fi
 
+if [ ! -f ~/.gvimrc ]
+then
+ln -s ~/.vim/.gvimrc ~/.gvimrc
+fi
+if [ ! -f ~/.vsvimrc ]
+then
+ln -s ~/.vim/.vsvimrc ~/.vsvimrc
+fi
+
+# make necessary directories
 if [ ! -d ~/.vim ]
 then
 mkdir ~/.vim
@@ -31,9 +42,10 @@ then
 mkdir ~/.vim/vimbackups/undos
 fi
 
+# make necessary files
 if [ ! -f ~/.vim/quick_locations.txt ]
 then
-printf '# this is a comment\nentries are of form "label":path\n\n# This File\n"Quick Locations":~\.vim\quick_locations.txt > ~/.vim/quick_locations.txt'
+printf '# this is a comment\nentries are of form "label":path\n\n# This File\n"Quick Locations":~\.vim\quick_locations.txt' > ~/.vim/quick_locations.txt
 fi
 
 if [ ! -f ~/.vim/vimrc_exceptions.vim ]
